moved all opcode test to proper location
